+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Stripe Payments Changelog
 
+## 1.1.4 - 2018.06.19
+### Fixed
+- Fixed issue with custom amounts
+
+## 1.1.3 - 2018.06.19
+### Added
+- Added support to currencies with zero decimals
+- Added link to view customer info in Stripe (Order view)
+
+### Fixed
+- Fixed issue where Frequency was not displayed after save
+- Fixed issue when Free trial Period was not set.
+
+### Added
+- Added Set Status element action to Orders
+
 ## 1.1.2 - 2018.06.10
 
 ### Added